Simplified paths with 1 file in a package now shows up as a file
If a package contained a single file, it would get "simplified" so
pkg/file will be one node. In this case both the file and pkg flags
would be set on the node causing it do render a package symbol instead
of a file icon. This has now been fixed.